# apps/web/openapi.codegen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sahabatkreator API (codegen)",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "summary": "Daftar akun sosial",
        "description": "Semua akun sosial milik organisasi (termasuk status koneksi).",
        "tags": ["Accounts"],
        "x-required-scopes": ["accounts:read"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daftar akun.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{id}/statistic": {
      "get": {
        "summary": "Statistik akun (bridge)",
        "description": "Statistik real-time satu akun via bridge Repliz. Akun non-bridge membalas 400/503.",
        "tags": ["Accounts"],
        "x-required-scopes": ["accounts:read"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID akun"
          }
        ],
        "responses": {
          "200": {
            "description": "Statistik akun.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountStatisticResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts": {
      "get": {
        "summary": "Daftar post",
        "description": "Daftar grup post (draft, terjadwal, sudah tayang) milik organisasi.",
        "tags": ["Posts"],
        "x-required-scopes": ["posts:read"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter YYYY-MM-DD"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter YYYY-MM-DD"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter status post"
          },
          {
            "name": "includeExternal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["1"]
            },
            "description": "Sertakan post eksternal"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostsListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Buat post",
        "description": "Membuat grup post baru. Bila `scheduledAt` diisi, menjadwalkannya (butuh fitur scheduled_posts).",
        "tags": ["Posts"],
        "x-required-scopes": ["posts:write"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Grup post dibuat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{id}": {
      "get": {
        "summary": "Detail post",
        "description": "",
        "tags": ["Posts"],
        "x-required-scopes": ["posts:read"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID grup post"
          }
        ],
        "responses": {
          "200": {
            "description": "Detail grup post + post items.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Hapus post",
        "description": "",
        "tags": ["Posts"],
        "x-required-scopes": ["posts:write"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID grup post"
          }
        ],
        "responses": {
          "200": {
            "description": "Post dihapus.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{id}/publish": {
      "post": {
        "summary": "Publish segera",
        "description": "Mem- enqueue publish untuk post yang bisa dipublish. Membalas `{ok,queued}` bila masuk queue, atau `{ok,published,processing,failed}` bila dikerjakan inline.",
        "tags": ["Posts"],
        "x-required-scopes": ["posts:write"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID grup post"
          }
        ],
        "responses": {
          "200": {
            "description": "Hasil publish.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PublishQueuedResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PublishInlineResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{id}/retry": {
      "post": {
        "summary": "Ulang publish post gagal",
        "description": "Hanya untuk post berstatus failed yang punya schedule bridge.",
        "tags": ["Posts"],
        "x-required-scopes": ["posts:write"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID grup post"
          }
        ],
        "responses": {
          "200": {
            "description": "Retry dijalankan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Key tidak memiliki scope yang dibutuhkan endpoint ini.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Bridge Repliz belum dikonfigurasi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
...
      "AccountsResponse": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountItem"
            }
          }
        },
        "required": ["accounts"]
      },
...
      "AccountStatisticResponse": {
        "type": "object",
        "properties": {
          "statistic": {
            "type": ["object", "null"],
            "properties": {
              "followers": {
                "type": ["number", "null"]
              },
              "posts": {
                "type": ["number", "null"]
              },
              "engagement": {
                "type": ["number", "null"]
              }
            },
            "required": ["followers", "posts", "engagement"],
            "description": "Statistik real-time dari bridge Repliz. null bila belum pernah disinkronkan."
          },
          "unsupported": {
            "type": "boolean"
          }
        },
        "required": ["statistic"]
      },
      "PostsListResponse": {
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostGroupItem"
            }
          },
          "page": {
            "type": "number"
          },
          "perPage": {
            "type": "number"
          }
        },
        "required": ["groups", "page", "perPage"]
      },
...
      "PostDetailResponse": {
        "type": "object",
        "properties": {
          "group": {
            "$ref": "#/components/schemas/PostGroupItem"
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostItem"
            }
          }
        },
        "required": ["group", "posts"]
      },
      "CreatePostRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "default": ""
          },
          "scheduledAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "timezone": {
            "type": "string",
            "default": "Asia/Jakarta"
          },
          "audioTrackId": {
            "type": ["string", "null"]
          },
          "productIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePostItem"
            },
            "minItems": 1
          }
        },
        "required": ["items"]
      },
...
      "CreatePostResponse": {
        "type": "object",
        "properties": {
          "postGroupId": {
            "type": "string"
          }
        },
        "required": ["postGroupId"]
      },
      "PublishQueuedResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [true]
          },
          "queued": {
            "type": "number"
          }
        },
        "required": ["ok", "queued"]
      },
...
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [true]
          }
        },
        "required": ["ok"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key format `sk_api_<32 karakter>`. Dibuat di Settings → API. Bisa juga dikirim via header `X-API-Key`."
      }
    }
  }
}